requirements.txt
PyJWT~=2.9.0
passlib~=1.7.4
fastapi~=0.115.0
SQLAlchemy~=2.0.35
alembic~=1.13.2